move validate_unique_hostnames to ValidateHost
