+++ lib/pharos/cluster_manager.rb
@@ -52,7 +52,6 @@ module Pharos
       addon_manager.validate
       gather_facts
       apply_phase(Phases::ValidateHost, config.hosts, ssh: true, parallel: true)
-      apply_phase(Phases::ValidateClusterHosts, config.hosts, ssh: false, parallel: false)
     end
 
     # @return [Array<Pharos::Configuration::Host>]
