+++ spec/pharos/phases/validate_host_spec.rb
@@ -108,4 +108,40 @@ describe Pharos::Phases::ValidateHost do
       end
     end
   end
+
+  describe '#validate_unique_hostnames' do
+    let(:config) { Pharos::Config.new(
+      hosts: [
+        Pharos::Configuration::Host.new(
+          address: '192.0.2.1',
+        ),
+        Pharos::Configuration::Host.new(
+          address: '192.0.2.2',
+        ),
+        Pharos::Configuration::Host.new(
+          address: '192.0.2.3',
+        )
+      ]
+    ) }
+
+    context 'no duplicate hostnames' do
+      it 'does not raise if no duplicates' do
+        config.hosts[0].hostname = "host-0"
+        config.hosts[1].hostname = "host-1"
+        config.hosts[2].hostname = "host-2"
+
+        subject.validate_unique_hostnames
+      end
+    end
+
+    context 'duplicate hostnames' do
+      it 'raises if duplicates' do
+        config.hosts[0].hostname = "foo"
+        config.hosts[1].hostname = "foo"
+        config.hosts[2].hostname = "foo"
+
+        expect{ subject.validate_unique_hostnames }.to raise_error(Pharos::InvalidHostError, "Duplicate hostname foo for hosts 192.0.2.2,192.0.2.3")
+      end
+    end
+  end
 end
